Correction majeur de tt Placement des speeches dans un dossier Deplacement des fonctions dans un file.py dedier Utilisation du os
