+++ speeches/Nomination_Chirac1.txt
