+++ speeches/Nomination_Chirac2.txt
