+++ speeches/Nomination_Giscard dEstaing.txt
