+++ speeches/Nomination_Hollande.txt
