+++ speeches/Nomination_Mitterrand1.txt
