+++ speeches/Nomination_Mitterrand2.txt
