+++ speeches/Nomination_Sarkozy.txt
